+++ src/advanced/atoms/isAdminAtom.ts
@@ -1,0 +1,3 @@
+import { atom } from 'jotai';
+
+export const isAdminAtom = atom<boolean>(false);
