+++ src/advanced/main.tsx
@@ -1,9 +1,13 @@
-import React from 'react'
-import ReactDOM from 'react-dom/client'
-import App from './App.tsx'
+import { Provider } from 'jotai';
+import React from 'react';
+import ReactDOM from 'react-dom/client';
+
+import App from './App.tsx';
 
 ReactDOM.createRoot(document.getElementById('root')!).render(
   <React.StrictMode>
-    <App />
-  </React.StrictMode>,
-)
+    <Provider>
+      <App />
+    </Provider>
+  </React.StrictMode>
+);
